Moved constant DEFAULT_GATEWAY_LANGUAGE to constants file
