+++ src/js/common/constants.js
@@ -29,6 +29,7 @@ export const tc_EDIT_VERSION_KEY = "tc_edit_version"; // do not change this stri
 export const tc_MIN_COMPATIBLE_VERSION_KEY = "tc_min_compatible_version"; // do not change this string as it will break compatibility with other versions
 export const tc_MIN_VERSION_ERROR = "TC_MIN_VERSION_ERROR";
 
+export const DEFAULT_GATEWAY_LANGUAGE = 'en';
 export const ORIGINAL_LANGUAGE = "originalLanguage";
 export const TARGET_LANGUAGE = "targetLanguage";
 export const TARGET_BIBLE = "targetBible";
